adding metric protocol w known mypy error
