+++ .changes/unreleased/Features-20230516-112250.yaml
@@ -1,0 +1,6 @@
+kind: Features
+body: Adding metric protocol
+time: 2023-05-16T11:22:50.448901-05:00
+custom:
+  Author: callum-mcdata
+  Issue: "18"
