+++ dbt_semantic_interfaces/objects/metric.py
@@ -11,7 +11,12 @@ from dbt_semantic_interfaces.objects.base import (
 )
 from dbt_semantic_interfaces.objects.filters.where_filter import WhereFilter
 from dbt_semantic_interfaces.objects.metadata import Metadata
-from dbt_semantic_interfaces.references import MeasureReference, MetricReference
+from dbt_semantic_interfaces.protocols.metric import (
+    _MetricInputMeasureMixin,
+    _MetricInputMixin,
+    _MetricMixin,
+    _MetricTypeParamsMixin,
+)
 from dbt_semantic_interfaces.type_enums.metric_type import MetricType
 from dbt_semantic_interfaces.type_enums.time_granularity import (
     TimeGranularity,
@@ -19,7 +24,7 @@ from dbt_semantic_interfaces.type_enums.time_granularity import (
 )
 
 
-class MetricInputMeasure(PydanticCustomInputParser, HashableBaseModel):
+class MetricInputMeasure(_MetricInputMeasureMixin, PydanticCustomInputParser, HashableBaseModel):
     """Provides a pointer to a measure along with metric-specific processing directives.
 
     If an alias is set, this will be used as the string name reference for this measure after the aggregation
@@ -45,16 +50,6 @@ class MetricInputMeasure(PydanticCustomInputParser, HashableBaseModel):
                 f"MetricInputMeasure inputs from model configs are expected to be of either type string or "
                 f"object (key/value pairs), but got type {type(input)} with value: {input}"
             )
-
-    @property
-    def measure_reference(self) -> MeasureReference:
-        """Property accessor to get the MeasureReference associated with this metric input measure."""
-        return MeasureReference(element_name=self.name)
-
-    @property
-    def post_aggregation_measure_reference(self) -> MeasureReference:
-        """Property accessor to get the MeasureReference with the aliased name, if appropriate."""
-        return MeasureReference(element_name=self.alias or self.name)
 
 
 class MetricTimeWindow(PydanticCustomInputParser, HashableBaseModel):
@@ -114,7 +109,7 @@ class MetricTimeWindow(PydanticCustomInputParser, HashableBaseModel):
         )
 
 
-class MetricInput(HashableBaseModel):
+class MetricInput(_MetricInputMixin, HashableBaseModel):
     """Provides a pointer to a metric along with the additional properties used on that metric."""
 
     name: str
@@ -123,12 +118,8 @@ class MetricInput(HashableBaseModel):
     offset_window: Optional[MetricTimeWindow]
     offset_to_grain: Optional[TimeGranularity]
 
-    @property
-    def as_reference(self) -> MetricReference:  # noqa: D
-        return MetricReference(element_name=self.name)
 
-
-class MetricTypeParams(HashableBaseModel):
+class MetricTypeParams(_MetricTypeParamsMixin, HashableBaseModel):
     """Type params add additional context to certain metric types (the context depends on the metric type)."""
 
     measure: Optional[MetricInputMeasure]
@@ -140,18 +131,8 @@ class MetricTypeParams(HashableBaseModel):
     grain_to_date: Optional[TimeGranularity]
     metrics: Optional[List[MetricInput]]
 
-    @property
-    def numerator_measure_reference(self) -> Optional[MeasureReference]:
-        """Return the measure reference, if any, associated with the metric input measure defined as the numerator."""
-        return self.numerator.measure_reference if self.numerator else None
 
-    @property
-    def denominator_measure_reference(self) -> Optional[MeasureReference]:
-        """Return the measure reference, if any, associated with the metric input measure defined as the denominator."""
-        return self.denominator.measure_reference if self.denominator else None
-
-
-class Metric(HashableBaseModel, ModelWithMetadataParsing):
+class Metric(_MetricMixin, HashableBaseModel, ModelWithMetadataParsing):
     """Describes a metric."""
 
     name: str
@@ -160,27 +141,3 @@ class Metric(HashableBaseModel, ModelWithMetadataParsing):
     type_params: MetricTypeParams
     filter: Optional[WhereFilter]
     metadata: Optional[Metadata]
-
-    @property
-    def input_measures(self) -> List[MetricInputMeasure]:
-        """Return the complete list of input measure configurations for this metric."""
-        tp = self.type_params
-        res = tp.measures or []
-        if tp.measure:
-            res.append(tp.measure)
-        if tp.numerator:
-            res.append(tp.numerator)
-        if tp.denominator:
-            res.append(tp.denominator)
-
-        return res
-
-    @property
-    def measure_references(self) -> List[MeasureReference]:
-        """Return the measure references associated with all input measure configurations for this metric."""
-        return [x.measure_reference for x in self.input_measures]
-
-    @property
-    def input_metrics(self) -> List[MetricInput]:
-        """Return the associated input metrics for this metric."""
-        return self.type_params.metrics or []
